workflow_logic/requirements.txt
# langchain
# langchain_openai
# langchain-community
# langgraph
# pytz
# stackapi
# xmltodict
# whisper
# tiktoken
asynctest
pytest
pytest-asyncio
pyautogen
pyautogen["anthropic"]
anthropic
exa_py
aiohttp
praw
arxiv
wikipedia
replicate
openai
pymongo
jinja2
python-dotenv
pillow
google-api-python-client
fastapi
pydantic
uvicorn
fastapi-cors